requirements.txt
openai-whisper==20231117
pyannote.audio==3.1.1
torch==2.1.0
torchaudio==2.1.0
transformers==4.35.2
librosa==0.10.1
soundfile==0.12.1
pydub==0.25.1
python-dotenv==1.0.0
click==8.1.7
tqdm==4.66.1
matplotlib==3.8.2
pandas==2.1.4
numpy==1.25.2
scipy==1.11.4